requirements.txt
aioboto3==13.4.0
aiobotocore==2.18.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aioitertools==0.12.0
aiosignal==1.3.2
aiosqlite==0.21.0
alembic==1.14.1
altair==5.5.0
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.8.0
arize-phoenix==8.8.0
arize-phoenix-client==1.0.3
arize-phoenix-evals==0.20.3
arize-phoenix-otel==0.8.0
asttokens==3.0.0
asyncer==0.0.7
attrs==25.1.0
Authlib==1.5.1
azure-core==1.32.0
azure-identity==1.20.0
banks==2.1.2
beautifulsoup4==4.13.3
bidict==0.23.1
blinker==1.9.0
boto3==1.36.1
botocore==1.36.1
branca==0.8.1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chainlit==2.2.1
charset-normalizer==3.4.1
chevron==0.14.0
click==8.1.8
cohere==5.14.0
colorama==0.4.6
contourpy==1.3.1
cryptography==44.0.1
cycler==0.12.1
dataclasses-json==0.6.7
decorator==5.2.1
Deprecated==1.2.18
deprecation==2.1.0
dirtyjson==1.0.8
distlib==0.3.9
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.16
dominate==2.9.1
duckdb==1.2.0
et_xmlfile==2.0.0
eval_type_backport==0.2.2
executing==2.2.0
fastapi==0.115.10
fastavro==1.10.0
filelock==3.17.0
filetype==1.2.0
Flask==3.1.0
Flask-Bootstrap==3.3.7.1
folium==0.19.5
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.2.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0rc0
google-api-python-client==2.168.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-genai==1.12.1
google-generativeai==0.8.5
googleapis-common-protos==1.68.0
graphql-core==3.2.6
greenlet==3.1.1
griffe==1.7.3
grpc-interceptor==0.15.4
grpcio==1.71.0
grpcio-status==1.71.0
grpclib==0.4.7
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.1
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.5.0
instructor==1.7.2
ipython==9.0.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpickle==4.0.2
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
lancedb==0.21.2
langchain==0.3.23
langchain-core==0.3.54
langchain-text-splitters==0.3.8
langsmith==0.3.32
Lazify==0.4.0
literalai==0.1.103
llama-cloud==0.1.13
llama-cloud-services==0.6.3
llama-index==0.12.22
llama-index-agent-openai==0.4.6
llama-index-cli==0.4.1
llama-index-core==0.12.33.post1
llama-index-embeddings-adapter==0.3.0
llama-index-embeddings-bedrock==0.5.0
llama-index-embeddings-google-genai==0.1.0
llama-index-embeddings-ollama==0.5.0
llama-index-embeddings-openai==0.3.1
llama-index-experimental==0.5.4
llama-index-finetuning==0.3.0
llama-index-indices-managed-llama-cloud==0.6.8
llama-index-llms-anthropic==0.6.7
llama-index-llms-azure-openai==0.3.0
llama-index-llms-bedrock==0.3.4
llama-index-llms-google-genai==0.1.7
llama-index-llms-mistralai==0.3.3
llama-index-llms-ollama==0.5.2
llama-index-llms-openai==0.3.25
llama-index-multi-modal-llms-openai==0.4.3
llama-index-postprocessor-cohere-rerank==0.3.0
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.5
llama-index-readers-llama-parse==0.4.0
llama-index-vector-stores-lancedb==0.3.2
llama-parse==0.6.2
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistralai==1.5.0
modal==0.74.14
mpmath==1.3.0
msal==1.31.1
msal-extensions==1.2.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.41.1
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
ollama==0.4.7
openai==1.65.2
openinference-instrumentation==0.1.23
openinference-semantic-conventions==0.1.14
openpyxl==3.1.5
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-exporter-otlp-proto-http==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
orjson==3.10.16
overrides==7.7.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pdfminer.six==20250327
pdfplumber==0.11.6
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
portalocker==2.10.1
prompt_toolkit==3.0.50
propcache==0.3.0
proto-plus==1.26.1
protobuf==5.29.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.10.1
pylance==0.25.2
pymongo==4.11.2
pyparsing==3.2.1
pypdf==5.3.0
PyPDF2==3.0.1
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-multipart==0.0.18
python-socketio==5.12.1
pytz==2025.1
pyvis==0.3.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.25.1
rsa==4.9
s3transfer==0.11.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
sentence-transformers==3.4.1
shellingham==1.5.4
sigtools==4.0.1
simple-websocket==1.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
sqlean.py==3.47.0
stack-data==0.6.3
starlette==0.41.3
strawberry-graphql==0.253.1
streamlit==1.45.1
striprtf==0.0.26
sympy==1.13.3
syncer==2.0.3
synchronicity==0.9.11
tantivy==0.22.2
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.9.0
tokenizers==0.21.0
toml==0.10.2
tomli==2.2.1
torch==2.2.2
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.49.0
typer==0.15.2
types-certifi==2021.10.8.3
types-requests==2.32.0.20250301
types-toml==0.10.8.20240310
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
uptrace==1.29.0
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
virtualenv==20.29.3
visitor==0.1.3
watchfiles==0.20.0
wcwidth==0.2.13
websockets==15.0
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
xyzservices==2025.1.0
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
